+++ libraries/core/test/search/metaheuristics/MOSA.test.ts
@@ -28,7 +28,7 @@ import { DummySearchSubject } from "../../mocks/DummySubject.mock";
 import { BranchObjectiveFunction } from "../../../lib";
 import { MockedMOSA } from "../../mocks/MOSAAdapter";
 import { DummyCrossover } from "../../mocks/DummyCrossover.mock";
-import { DummyProcreation } from "../../mocks/DummyOffspring.mock";
+import { DummyProcreation } from "../../mocks/DummyProcreation.mock";
 import { DummyBranchDistance } from "../../mocks/DummyBranchDistance.mock";
 
 const expect = chai.expect;
@@ -72,7 +72,7 @@ describe("Test MOSA", function () {
     const mockedRunner = <EncodingRunner<DummyEncodingMock>>{};
     const mockedSampler = <EncodingSampler<DummyEncodingMock>>{};
     const mockedCrossover = new DummyCrossover(0.8, 0.8);
-    const mockedOffspring = new DummyProcreation(
+    const mockedProcreation = new DummyProcreation(
       mockedCrossover,
       mockedSampler
     );
@@ -80,7 +80,7 @@ describe("Test MOSA", function () {
     const mosa = new MOSAFamily(
       new UncoveredObjectiveManager(mockedRunner, new Set()),
       mockedSampler,
-      mockedOffspring,
+      mockedProcreation,
       50
     );
     const frontZero = mosa.preferenceCriterion(
@@ -112,7 +112,7 @@ describe("Test MOSA", function () {
     const mockedRunner = <EncodingRunner<DummyEncodingMock>>{};
     const mockedSampler = <EncodingSampler<DummyEncodingMock>>{};
     const mockedCrossover = new DummyCrossover(0.8, 0.8);
-    const mockedOffspring = new DummyProcreation(
+    const mockedProcreation = new DummyProcreation(
       mockedCrossover,
       mockedSampler
     );
@@ -120,7 +120,7 @@ describe("Test MOSA", function () {
     const mosa = new MOSAFamily(
       new UncoveredObjectiveManager(mockedRunner, new Set()),
       mockedSampler,
-      mockedOffspring,
+      mockedProcreation,
       50
     );
     const front = mosa.getNonDominatedFront(objectives, [
@@ -153,7 +153,7 @@ describe("Test MOSA", function () {
     const mockedRunner = <EncodingRunner<DummyEncodingMock>>{};
     const mockedSampler = <EncodingSampler<DummyEncodingMock>>{};
     const mockedCrossover = new DummyCrossover(0.8, 0.8);
-    const mockedOffspring = new DummyProcreation(
+    const mockedProcreation = new DummyProcreation(
       mockedCrossover,
       mockedSampler
     );
@@ -161,7 +161,7 @@ describe("Test MOSA", function () {
     const mosa = new MOSAFamily(
       new UncoveredObjectiveManager(mockedRunner, new Set()),
       mockedSampler,
-      mockedOffspring,
+      mockedProcreation,
       4
     );
     const front = mosa.preferenceSortingAlgorithm(
@@ -199,7 +199,7 @@ describe("Test MOSA", function () {
     const mockedRunner = <EncodingRunner<DummyEncodingMock>>{};
     const mockedSampler = <EncodingSampler<DummyEncodingMock>>{};
     const mockedCrossover = new DummyCrossover(0.8, 0.8);
-    const mockedOffspring = new DummyProcreation(
+    const mockedProcreation = new DummyProcreation(
       mockedCrossover,
       mockedSampler
     );
@@ -207,7 +207,7 @@ describe("Test MOSA", function () {
     const mosa = new MockedMOSA(
       new UncoveredObjectiveManager(mockedRunner, new Set()),
       mockedSampler,
-      mockedOffspring,
+      mockedProcreation,
       50
     );