+++ plugins/plugin-core-sfuzz/lib/SFuzzModule.ts
@@ -17,14 +17,14 @@
  */
 
 import { Module, ModuleManager } from "@syntest/module";
-import { SfuzzPlugin } from "./plugins/SfuzzPlugin";
-import { SfuzzObjectiveManagerPlugin } from "./plugins/SfuzzObjectiveManagerPlugin";
-import SfuzzPreset from "./SfuzzPreset";
+import { SFuzzPlugin } from "./plugins/SFuzzPlugin";
+import { SFuzzObjectiveManagerPlugin } from "./plugins/SFuzzObjectiveManagerPlugin";
+import { SFuzzPreset } from "./presets/_SFuzzPreset";
 
-export default class SfuzzModule extends Module {
+export default class SFuzzModule extends Module {
   async register(moduleManager: ModuleManager): Promise<void> {
-    moduleManager.registerPlugin(this.name, new SfuzzObjectiveManagerPlugin());
-    moduleManager.registerPlugin(this.name, new SfuzzPlugin());
-    moduleManager.registerPreset(this.name, new SfuzzPreset());
+    moduleManager.registerPlugin(this.name, new SFuzzObjectiveManagerPlugin());
+    moduleManager.registerPlugin(this.name, new SFuzzPlugin());
+    moduleManager.registerPreset(this.name, new SFuzzPreset());
   }
 }