+++ libraries/core/CHANGELOG.md
@@ -9,7 +9,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - add event emissions in search algorithm and target pool ([#201](https://github.com/syntest-framework/syntest-core/issues/201)) ([c5e499a](https://github.com/syntest-framework/syntest-core/commit/c5e499af53097b6881416528d914795f67ab541d))
 - change changelog preset ([#198](https://github.com/syntest-framework/syntest-core/issues/198)) ([01df511](https://github.com/syntest-framework/syntest-core/commit/01df511a936cce6851259a512b6ea70760ad8dd4))
-- Extract Sfuzz as a plugin
+- Extract sFuzz as a plugin
 - Extract graphing related code as plugin
 - Extract CFG related code as a library
 - abstract Launcher class with common functions